update constrain tests
